+++ test/notifier.js
@@ -46,25 +46,25 @@ vows.describe('notifier pending requests').addBatch({
   }
 }).export(module, {error: false});
 
-var shutdownSandbox = null;
-vows.describe('notifier shutdown').addBatch({
+var waitSandbox = null;
+vows.describe('notifier wait').addBatch({
   'has pending requests': {
     topic: function() {
       assert(notifier.pendingItemsCount() == 0);
-      shutdownSandbox = sinon.sandbox.create();
-      shutdownSandbox.stub(https, 'request', function(){ /* do nothing */ });
+      waitSandbox = sinon.sandbox.create();
+      waitSandbox.stub(https, 'request', function(){ /* do nothing */ });
 
       notifier.handleError(new Error('test'));
       assert(notifier.pendingItemsCount() == 1); // should have been enqueued synchronously
 
-      notifier.shutdown(this.callback);
+      notifier.wait(this.callback);
 
       notifier.handleError(new Error('test'));
       assert(notifier.pendingItemsCount() == 2); // should have been enqueued synchronously
     },
     'it keeps track of pending requests': function() {
       assert(notifier.pendingItemsCount() == 0); // callback should be called after all items are flushed
-      shutdownSandbox.restore();
+      waitSandbox.restore();
     }
   }
 }).export(module, {error: false});